refactor: made clientService functions more descriptive
